+++ src/main/java/com/lowdragmc/lowdraglib2/gui/texture/TransformTexture.java
@@ -68,13 +68,13 @@ public abstract class TransformTexture implements IGuiTexture {
 
     @Override
     @OnlyIn(Dist.CLIENT)
-    public final void draw(GuiGraphics graphics, int mouseX, int mouseY, float x, float y, float width, float height, float partialTicks) {
+    public final void draw(GuiGraphics graphics, float mouseX, float mouseY, float x, float y, float width, float height, float partialTicks) {
         preDraw(graphics, x, y, width, height);
         drawInternal(graphics, mouseX, mouseY, x, y, width, height, partialTicks);
         postDraw(graphics, x, y, width, height);
     }
 
     @OnlyIn(Dist.CLIENT)
-    protected abstract void drawInternal(GuiGraphics graphics, int mouseX, int mouseY, float x, float y, float width, float height, float partialTicks);
+    protected abstract void drawInternal(GuiGraphics graphics, float mouseX, float mouseY, float x, float y, float width, float height, float partialTicks);
 
 }
